Add Tree algorithm
Refactoring of number of elements types.
Refactor Reducers, let them do 3 types of buffer
compression mode:
fused, non-fused, non-fused per entry.

+++ horovod/common/ops/compressed/compression/cuda/cuda_arithmetic_functions.h
@@ -7,7 +7,7 @@ namespace common {
 namespace cuda {
 
 template <typename T>
-void CUDA_add(int64_t n, const T* x, T* y, T* sum, cudaStream_t stream);
+void CUDA_add(int n, const T* x, T* y, T* sum, cudaStream_t stream);
 
 } // namespace cuda
 } // namespace common
